重构文章列表页面
