+++ app/models/article/crud.go
@@ -15,3 +15,12 @@ func Get(idstr string) (Article, error) {
 
 	return article, nil
 }
+
+// GetAll 显示所有的文章
+func GetAll() ([]Article, error) {
+	var articles []Article
+	if err := model.DB.Find(&articles).Error; err != nil {
+		return articles, err
+	}
+	return articles, nil
+}
