+++ routes/web.go
@@ -19,4 +19,7 @@ func RegisterWebRoutes(r *mux.Router) {
 	//文章相关页面
 	ac := new(controllers.ArticlesController)
 	r.HandleFunc("/articles/{id:[0-9]+}", ac.Show).Methods("GET").Name("articles.show")
+
+	//显示所有的文章
+	r.HandleFunc("/articles", ac.Index).Methods("GET").Name("articles.index")
 }
